Move configureCorporateProxyAgent to a separate file
This would simplify the merging process in case of upstream changes
